requirements.txt
click
luigi
nose
m2r
PyYAML
